requirements.txt
fastapi
uvicorn
openai
google-genai
python-dotenv
requests
python-multipart 